+++ lib/src/errors/dependency_injection_error.dart
@@ -2,4 +2,9 @@ class DependencyInjectionError extends Error {
   DependencyInjectionError(this.message);
 
   final String message;
+
+  @override
+  String toString() {
+    return 'DependencyInjectionError: $message';
+  }
 }